pyproject.toml
[project]
name = "g0253"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.12"
dependencies = [
    "asyncio>=4.0.0",
    "cryptography>=46.0.5",
    "ecdsa>=0.19.1",
    "polars>=1.39.0",
    "xrpl-py>=4.5.0",
]
